:art: rename module
