+++ app/tests/e2e/issuer/ld_proof/test_ld_proof_did_sov.py
@@ -14,7 +14,7 @@ from app.models.issuer import SendCredential
 from app.routes.issuer import router as issuer_router
 from app.routes.oob import router as oob_router
 from app.routes.wallet.dids import router as wallet_router
-from app.tests.util.ecosystem_connections import FaberAliceConnect
+from app.tests.util.member_connections import FaberAliceConnect
 from app.tests.util.webhooks import check_webhook_state
 from shared import RichAsyncClient
