+++ app/tests/e2e/issuer/test_save_exchange_record.py
@@ -7,7 +7,7 @@ from fastapi import HTTPException
 
 from app.routes.issuer import router
 from app.tests.util.credentials import sample_credential_attributes
-from app.tests.util.ecosystem_connections import FaberAliceConnect
+from app.tests.util.member_connections import FaberAliceConnect
 from app.tests.util.webhooks import check_webhook_state
 from shared import RichAsyncClient
 from shared.models.credential_exchange import CredentialExchange
