+++ app/tests/e2e/test_connections.py
@@ -6,7 +6,7 @@ from fastapi import HTTPException
 
 from app.models.connections import AcceptInvitation, CreateInvitation
 from app.routes.connections import router
-from app.tests.util.ecosystem_connections import BobAliceConnect
+from app.tests.util.member_connections import BobAliceConnect
 from app.tests.util.webhooks import check_webhook_state
 from shared import RichAsyncClient
