+++ app/tests/e2e/test_messaging.py
@@ -3,7 +3,7 @@ from assertpy.assertpy import assert_that
 
 from app.models.messaging import Message, TrustPingMsg
 from app.routes.messaging import router
-from app.tests.util.ecosystem_connections import BobAliceConnect
+from app.tests.util.member_connections import BobAliceConnect
 from shared import RichAsyncClient
 
 MESSAGING_BASE_PATH = router.prefix
