+++ webhooks/tests/e2e/test_sse.py
@@ -4,7 +4,7 @@ import logging
 import pytest
 from httpx import ReadTimeout, Response, Timeout
 
-from app.tests.util.ecosystem_connections import BobAliceConnect
+from app.tests.util.member_connections import BobAliceConnect
 from app.tests.util.webhooks import get_wallet_id_from_async_client
 from shared import WEBHOOKS_URL, RichAsyncClient
